app/requirements.txt
fastapi==0.115.4
mlflow==2.17.0
nltk==3.9.1
numpy==2.1.2
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn==0.32.0
scikit-learn==1.5.2
lightgbm==4.5.0
